added statistics accumulators, added NC incremental, Flink 1.5, refactoring
- added statistics method to retrieve cluster sizes from graphs/vertices
- added quality method for NC dataset
- switched name/surname for NC when reading Gradoop graphs
- removed (some) old clustering strategies
- removed unneeded DefaultProcessing constructors, refactoring
- added maven findbugs and jar version
- updated flink-gelly, flink-clients dependency
- updated to Flink 1.5.0

+++ analysis-core/src/main/java/org/mappinganalysis/model/functions/decomposition/simsort/SimSortComputeFunction.java
@@ -51,7 +51,9 @@ class SimSortComputeFunction
         iterationAggSim += message.getEdgeSim();
       }
 
-      iterationAggSim = Utils.getExactDoubleResult(iterationAggSim, messageCount);
+      iterationAggSim = Utils.getExactDoubleResult(
+          iterationAggSim,
+          (double) messageCount);
 //      LOG.debug(vertex.getId() + " itAggSim: " + iterationAggSim + " old: " + vertex.getValue().getSim());
 
       if (Doubles.compare(vertexAggSim, Constants.DEFAULT_VERTEX_SIM) != 0
